--- a/usaspending_api/references/v2/views/autocomplete.py
+++ b/usaspending_api/references/v2/views/autocomplete.py
@@ -1,4 +1,4 @@
-from django.db.models import F
+from django.db.models import F, Q
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_extensions.cache.decorators import cache_response
@@ -40,11 +40,13 @@ class BaseAutocompleteViewSet(APIView):
     # Shared autocomplete...
     def agency_autocomplete(self, request):
         """Search by subtier agencies, return all, with toptiers first"""
+
         search_text, limit = self.get_request_payload(request)
 
         queryset = Agency.objects.filter(
-            subtier_agency__name__icontains=search_text). \
-            order_by('-toptier_flag')
+            Q(subtier_agency__name__icontains=search_text)
+            | Q(subtier_agency__abbreviation__icontains=search_text)
+            ).order_by('-toptier_flag')
 
         return Response(
             {'results': AgencySerializer(queryset[:limit], many=True).data}