Per bug DS-1708: 

> As identified in the PMO team's QA review: When searching in Funding Agency or Awarding Agency, "USDA" will not bring up Department of Agriculture, HHS does not bring up Health and Human Services, etc. If you put a space between each letter in the acronym it does work, but that is not intuitive to do.

https://federal-spending-transparency.atlassian.net/browse/DS-1708

```
url = 'http://localhost:8000/api/v2/autocomplete/awarding_agency'
payload = json.loads('''{"search_text":"usda","limit":20}''')
import requests
resp = requests.post(url, payload)
resp.json()
```
result 

```
{'results': [{'id': 112,
   'office_agency': None,
   'subtier_agency': {'abbreviation': 'USDA',
    'name': 'Department of Agriculture',
    'subtier_code': '1200'},
   'toptier_agency': {'abbreviation': 'USDA',
    'cgac_code': '012',
    'fpds_code': '1200',
    'name': 'Department of Agriculture'},
   'toptier_flag': True}]}
```

## Performance impact 

(testing on the chunk of code above with %timeit):

Without change, 667 ms ± 18 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

With change, 843 ms ± 112 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)